Rename python package to django_bridge
