+++ packages/create-django-render/template/server/__projectname__/templates/django_bridge/bootstrap.html
@@ -1,4 +1,4 @@
-{% extends "django_render/bootstrap.html" %}
+{% extends "django_bridge/bootstrap.html" %}
 
 {% comment %}
 Use this file to override the default bootstrap.html template.
