+++ python/.isort.cfg
@@ -1,3 +1,3 @@
 [settings]
-known_first_party = django_render
+known_first_party = django_bridge
 profile = black
