+++ python/django_bridge/adapters/__init__.py
