+++ python/django_bridge/adapters/registry.py
