+++ python/django_bridge/templates/django_bridge/bootstrap.html
