+++ python/django_bridge/test/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "django_bridge.test.apps.DjangoBridgeTestAppConfig"
