+++ python/django_bridge/test/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 
 
 class DjangoBridgeTestAppConfig(AppConfig):
-    label = "django_render_test"
-    name = "django_render.test"
+    label = "django_bridge_test"
+    name = "django_bridge.test"
     verbose_name = "Django Bridge tests"
