+++ python/django_bridge/views.py
