--- python/django_render/test/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = "django_render.test.apps.DjangoBridgeTestAppConfig"
